refactor(test): replace AbstractLocalStackTest with AbstractFlociTest using testcontainers-floci
Replace manual GenericContainer setup with FlociContainer from the
io.floci:testcontainers-floci:2.8.0 library, which provides a typed
API (getEndpoint(), getAccessKey(), getSecretKey()) and removes the
low-level port-binding boilerplate.
