+++ build.gradle
@@ -131,6 +131,7 @@ dependencies {
     // testcontainers
     testImplementation "org.testcontainers:testcontainers:2.0.5"
     testImplementation "org.testcontainers:junit-jupiter:1.21.4"
+    testImplementation "io.floci:testcontainers-floci:2.8.0"
 }
 
 /**********************************************************************************************************************\
