+++ src/test/java/io/kestra/plugin/aws/cloudformation/CreateTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
-import io.kestra.plugin.aws.AbstractLocalStackTest;
+import io.kestra.plugin.aws.AbstractFlociTest;
 
 import jakarta.inject.Inject;
 
@@ -15,7 +15,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
-class CreateTest extends AbstractLocalStackTest {
+class CreateTest extends AbstractFlociTest {
 
     @Inject
     private RunContextFactory runContextFactory;
