+++ src/test/java/io/kestra/plugin/aws/cloudformation/DeleteTest.java
@@ -7,11 +7,11 @@ import org.junit.jupiter.api.Test;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
-import io.kestra.plugin.aws.AbstractLocalStackTest;
+import io.kestra.plugin.aws.AbstractFlociTest;
 
 import jakarta.inject.Inject;
 
-class DeleteTest extends AbstractLocalStackTest {
+class DeleteTest extends AbstractFlociTest {
 
     @Inject
     private RunContextFactory runContextFactory;
