+++ src/test/java/io/kestra/plugin/aws/sns/AbstractSnsTest.java
@@ -4,7 +4,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 import io.kestra.core.junit.annotations.KestraTest;
 import io.kestra.core.runners.RunContextFactory;
-import io.kestra.plugin.aws.AbstractLocalStackTest;
+import io.kestra.plugin.aws.AbstractFlociTest;
 
 import jakarta.inject.Inject;
 import software.amazon.awssdk.services.sns.SnsClient;
@@ -12,7 +12,7 @@ import software.amazon.awssdk.services.sns.model.CreateTopicRequest;
 
 @KestraTest
 @Testcontainers
-public class AbstractSnsTest extends AbstractLocalStackTest {
+public class AbstractSnsTest extends AbstractFlociTest {
     static final String TOPIC_ARN = "arn:aws:sns:us-east-1:000000000000:test-topic";
 
     @Inject
